Move CLI workflow path filter into shared/workflow
Relocate src/cli/src/shared/fs/path-filter.js to src/cli/src/shared/workflow/path-filter.js so the workflow-specific CLI helpers sit beside fixture roots instead of the generic fs facade. Update dependent imports and tests to reflect the new home.
