+++ src/cli/src/shared/workflow/fixture-roots.js
@@ -2,7 +2,7 @@ import path from "node:path";
 
 import { toArray } from "../dependencies.js";
 import { REPO_ROOT } from "../workspace-paths.js";
-import { createWorkflowPathFilter } from "../fs/path-filter.js";
+import { createWorkflowPathFilter } from "./path-filter.js";
 
 export const DEFAULT_FIXTURE_DIRECTORIES = Object.freeze([
     path.resolve(REPO_ROOT, "src", "parser", "test", "input"),
