+++ src/cli/test/shared-workflow-path-filter.test.js
@@ -5,7 +5,7 @@ import { describe, it } from "node:test";
 import {
     createWorkflowPathFilter,
     ensureWorkflowPathsAllowed
-} from "../src/shared/fs/path-filter.js";
+} from "../src/shared/workflow/path-filter.js";
 
 describe("workflow path filter helpers", () => {
     it("allows paths that satisfy the workflow filters", () => {
